move database.js
